Moves video label tests
