+++ packages/video-label/package.json
@@ -8,9 +8,9 @@
     "fmt": "prettier --write '**/*.*'",
     "prettier:diff": "prettier --list-different '**/*.*'",
     "lint": "eslint . && yarn prettier:diff",
-    "test:android": "jest --config='./src/__tests__/android/jest.config.js'",
-    "test:ios": "jest --config='./src/__tests__/ios/jest.config.js'",
-    "test:web": "jest --config='./src/__tests__/web/jest.config.js'",
+    "test:android": "jest --config='./__tests__/android/jest.config.js'",
+    "test:ios": "jest --config='./__tests__/ios/jest.config.js'",
+    "test:web": "jest --config='./__tests__/web/jest.config.js'",
     "test:all": "yarn test:android --coverage & yarn test:ios --coverage & yarn test:web --coverage",
     "prepublishOnly": "yarn transpile",
     "transpile": "babel src --ignore node_modules,__tests__ -d dist"
